feat(drawer): put Drawer() class to tool.py as DrawerTool()
