+++ tiny_scientist/__init__.py
@@ -1,8 +1,7 @@
 from .coder import Coder
-from .drawer import Drawer
 from .reviewer import Reviewer
 from .scientist import TinyScientist
 from .thinker import Thinker
 from .writer import Writer
 
-__all__ = ["Coder", "Drawer", "Reviewer", "Thinker", "Writer", "TinyScientist"]
+__all__ = ["Coder", "Reviewer", "Thinker", "Writer", "TinyScientist"]
